Remove fixtures/mgmt_system.py (#4125)
