+++ conftest.py
@@ -180,7 +180,6 @@ pytest_plugins = (
     'fixtures.log',
     'fixtures.maximized',
     'fixtures.merkyl',
-    'fixtures.mgmt_system',
     'fixtures.nelson',
     'fixtures.node_annotate',
     'fixtures.page_screenshots',
